[graphics] add Sprite structure

+++ graphics/position.go
@@ -1,6 +1,6 @@
 package graphics
 
 type Position struct {
-	X float64 `json:"x"`
-	Y float64 `json:"y"`
+	X int `json:"x"`
+	Y int `json:"y"`
 }
